Add files via upload
This is the final project that I completed for my C++ course. The text file contains units and factors for 9 different types of engineering units such as power, weight, flow rate, pressure etc. The program reads the contents of the file into a one single dimensional array and two two dimensional arrays. It then asks for a number including units that the user wants to have converted. It recognizes the type of unit based on the unit itself. After receiving input from the user, it asks (in list form) what unit the user would like to convert the number to. After asking, the program converts the number in question to the requested unit such as mph to m/s. If the user submits a unit not included in the file, the program goes through a series of questions regarding the type of unit etc. The program also has a password protected method of adding additional capability. For instance, if the user submits a unit not included in the file, with the password, the user can have the unit and factor added